+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 # Changelog
+
 Changelog to track updates for this mod.  
-    Add your changes to Unreleased if you want to commit.  
-    Please write according to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
+Add your changes to Unreleased if you want to commit.  
+Please write according to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
 ### Added
+
+- Add cross block state provider
 
 ### Changed
 
@@ -20,34 +23,45 @@ Changelog to track updates for this mod.
 ## [3.4.0] - 2023-01-25
 
 ### Added
+
 - Add living tick event
 
 ## [3.3.0] - 2023-01-24
 
 ### Added
+
 - Add new utility class
 
 ## [3.3.0-beta.1] - 2023-01-22
 
 ### Changed
+
 - Port MC1.19.3
 
 ### Removed
+
 - Separated Fabric's built-in OBJ loader ([Special Model Loader](https://github.com/TeamFelnull/SpecialModelLoader))
 
 ### Fixed
+
 - Fixed a problem that crashes without notifying when Architecture is not installed (Forge Only)
 
 ## [3.3.0-alpha.1] - 2023-01-18
 
 ### Changed
+
 - Port MC1.19.3
 
 ### Removed
+
 - Separated Fabric's built-in OBJ loader ([Special Model Loader](https://github.com/TeamFelnull/SpecialModelLoader))
 
 [Unreleased]: https://github.com/TeamFelnull/OtyacraftEngine/compare/v3.4.0...HEAD
+
 [3.4.0]: https://github.com/TeamFelnull/OtyacraftEngine/compare/v3.3.0...v3.4.0
+
 [3.3.0]: https://github.com/TeamFelnull/OtyacraftEngine/compare/v3.3.0-beta.1...v3.3.0
+
 [3.3.0-alpha.1]: https://github.com/TeamFelnull/OtyacraftEngine/commits/v3.3.0-alpha.1
+
 [3.3.0-beta.1]: https://github.com/TeamFelnull/OtyacraftEngine/compare/v3.3.0-alpha.1...v3.3.0-beta.1